requirements.txt
pandas>=2.0.0
google-api-python-client>=2.115.0
google-auth>=2.22.0
google-auth-httplib2>=0.1.0
httplib2>=0.22.0
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
